Changes router from json to Route components.
Creates StateContext in App.jsx to manage the user login state
throughout the app.

Creates components for each link in the courses dropdownmenu in the
NavBar.

Creates a component "HomeMainBody" for the body of the homepage.

Makes the logged in user's firstName and lastName shown in the NavBar.

Creates logout button inside NavBar, which is available when user is
logged in.

Changes NavBar styling.

Other minor changes.